fix(p24): resolve the 25 real Psalm errors from the full-repo run
Now that psalm.xml can parse (bdb004655), a real full-repo run surfaced 25
errors -- 22 are genuine type/logic gaps, fixed directly rather than
suppressed:

- PhotosAddDirectPageRenderer/PictureFormatsPageRenderer/
  PictureModifyPageRenderer: dividing a ?int filesize by a float literal
  (1024.0) tripped strictBinaryOperands -- explicit (float) cast.
- PictureCoiPageRenderer (4x): `is_string($row['coi'] ?? null) ? $row['coi']
  : ''` -- Psalm can't correlate the guard with a differently-shaped branch
  expression. Extracted to a local variable first, then narrowed.
- ApiKeyService::get(): the loop overwrote $api_keys[$i] (a list of ApiKey
  objects) in place with plain arrays, so Psalm inferred an unprovable
  ApiKey|array union. Built a separate $results array instead.
- CategoryDefaultRenderer: concatenating a float|null rating score into a
  string without a cast.
- CommentsController (3x): $elements is genuinely keyed by
  ImageRepository::findByIds()'s int image ids, but the loop indexed it
  with a string -- worked at runtime only because PHP auto-coerces
  numeric-string array keys. Added a real int key for the lookup, kept the
  string $image_id for its other genuine string uses (URL params, template
  vars).
- PwgImages: `in_array($image_row[$key], [null, false, 0, 0.0, '0', '',
  []], true)` against a value whose real type is int|null|string --
  false/0.0/[] can never match. Narrowed the haystack to what can actually
  occur.
- BrowserTestHelpers::imagePath(): same guard/branch-expression mismatch as
  PictureCoiPageRenderer above.

3 are genuine Psalm/PHPStan disagreements or deliberate internal-API use,
documented in psalm.xml rather than worked around in code:
- PictureController.php: tried a `?? ''` fallback first -- it broke
  PHPStan, which already proves format_id is exhaustively present in that
  branch (nullCoalesce.offset). Reverted the code change, documented the
  disagreement instead.
- BrowserTestHelpers::clickWithTimeout(): deliberate use of Pest's internal
  GuessLocator (already justified in its own docblock, already has a
  matching @phpstan-ignore) -- added the mirroring Psalm suppression.
- DbConnectionTest.php: the same by-ref beforeEach/afterEach closure
  narrowing gap already documented for its sibling ConfigLoaderTest.php/
  DbCredentialsTest.php, just not yet added to that list.

Verified: full-repo phpstan/ecs/psalm all clean (0 errors, no unused
suppressions); Unit 590/590, Arch 30/30, Browser 70/70.

+++ src/Piwigo/Admin/PhotosAddDirectPageRenderer.php
@@ -185,7 +185,7 @@ SELECT COUNT(*)
 
                     foreach ($formats as $format) {
                         $format_ext = $format->ext;
-                        $format_filesize = ($format->filesize ?? 0) / 1024.0;
+                        $format_filesize = ((float) ($format->filesize ?? 0)) / 1024.0;
                         $format_strings[] = sprintf('%s (%.2fMB)', $format_ext, $format_filesize);
                         $formats_exts[] = strtolower($format_ext);
                     }
